improved code organisation
