+++ nengo_gui/static/components/2d_axes.js
