+++ nengo_gui/static/components/htmlview.js
