+++ nengo_gui/static/components/netgraph.js
