+++ nengo_gui/static/components/netgraph_conn.js
