+++ nengo_gui/static/components/netgraph_item.js
