+++ nengo_gui/static/components/pointer.js
