+++ nengo_gui/static/components/raster.js
