+++ nengo_gui/static/components/slider.css
