+++ nengo_gui/static/components/slider.js
