+++ nengo_gui/static/components/slidercontrol.js
