+++ nengo_gui/static/components/value.css
