+++ nengo_gui/static/components/value.js
