+++ nengo_gui/static/components/xyvalue.js
