+++ nengo_gui/templates/page.html
@@ -2,15 +2,15 @@
     <head>
         <link href="static/main.css" rel="stylesheet" type="text/css" />
         <link href="static/nengo.css" rel="stylesheet" type="text/css" />
-        <link href="static/value.css" rel="stylesheet" type="text/css" />
         <link href="static/menu.css" rel="stylesheet" type="text/css" />
-        <link href="static/raster.css" rel="stylesheet" type="text/css" />
-        <link href="static/pointer.css" rel="stylesheet" type="text/css" />
-        <link href="static/slider.css" rel="stylesheet" type="text/css" />
+        <link href="static/components/value.css" rel="stylesheet" type="text/css" />
+        <link href="static/components/raster.css" rel="stylesheet" type="text/css" />
+        <link href="static/components/pointer.css" rel="stylesheet" type="text/css" />
+        <link href="static/components/slider.css" rel="stylesheet" type="text/css" />
+        <link href="static/components/netgraph.css" rel="stylesheet" type="text/css" />
         <link href="static/modal.css" rel="stylesheet" type="text/css" />
         <link href="static/tooltips.css" rel="stylesheet" type="text/css" />
         <link href="static/sim_control.css" rel="stylesheet" type="text/css" />
-        <link href="static/netgraph.css" rel="stylesheet" type="text/css" />
         <link href="static/ace.css" rel="stylesheet" type="text/css" />
         <link href="static//lib/css/bootstrap.min.css" rel="stylesheet" type="text/css" />
         <link href="static/top_toolbar.css" rel="stylesheet" type="text/css" />
@@ -115,19 +115,21 @@
         <script src="static/lib/js/bootstrap.min.js"></script>
         <script src="static/lib/js/validator.js"></script>
         <script src="static/nengo.js"></script>
-        <script src="static/2d_axes.js"></script>
-        <script src="static/time_axes.js"></script>
-        <script src="static/value.js"></script>
-        <script src="static/xyvalue.js"></script>
-        <script src="static/raster.js"></script>
-        <script src="static/slider.js"></script>
-        <script src="static/slidercontrol.js"></script>
+        <script src="static/datastore.js"></script>
+        <script src="static/components/component.js"></script>
+        <script src="static/components/2d_axes.js"></script>
+        <script src="static/components/time_axes.js"></script>
+        <script src="static/components/value.js"></script>
+        <script src="static/components/xyvalue.js"></script>
+        <script src="static/components/raster.js"></script>
+        <script src="static/components/slider.js"></script>
+        <script src="static/components/slidercontrol.js"></script>
+        <script src="static/components/pointer.js"></script>
+        <script src="static/components/htmlview.js"></script>
+        <script src="static/components/netgraph.js"></script>
+        <script src="static/components/netgraph_item.js"></script>
+        <script src="static/components/netgraph_conn.js"></script>
         <script src="static/sim_control.js"></script>
-        <script src="static/netgraph.js"></script>
-        <script src="static/pointer.js"></script>
-        <script src="static/htmlview.js"></script>
-        <script src="static/netgraph_item.js"></script>
-        <script src="static/netgraph_conn.js"></script>
         <script src="static/top_toolbar.js"></script>
         <script src="static/modal.js"></script>
         <script src="static/tooltips.js"></script>
